requirements.txt
python-dotenv
selenium
parse
w3lib
multiprocessing
time
regex
datetime
random